refactored the handleChange to a utility function updateFormFields
