+++ src/shared-components/inventory/InventoryForm.tsx
@@ -4,6 +4,7 @@ import type { Database } from "../../lib/database.types";
 import deleteEntry from "../../utils.ts/inventory/deleteEntry";
 import { useInventoryForm } from "../../hooks/useInventoryForm";
 import addEntry from "../../utils.ts/inventory/addEntry";
+import updateFormFields from "./updateFormFields";
 
 // This pulls the exact row definition from your 'inventory' table
 type InventoryEntry = Database["public"]["Tables"]["inventory"]["Row"];
@@ -124,29 +125,6 @@ export function InventoryForm({
   ) => {
     await addEntry(e, formData, setAddInventory);
     setFormData(initialState);
-  };
-
-  const handleChange = (name: string, value: string) => {
-    setFormData((prev) => {
-      // Start with the new value
-      const updatedData = { ...prev, [name]: value };
-      const costPer = parseFloat(updatedData.cost_per_item as string) || 0;
-      const qty = parseInt(updatedData.quantity as string) || 0;
-      const listPrice = parseFloat(updatedData.sold_price as string) || 0;
-      const shipping = parseFloat(updatedData.shipping_cost as string) || 0;
-      const calculatedTotalCost = costPer * qty;
-      const calculatedTotalPriceSold = listPrice * qty;
-      const calculatedProfit =
-        calculatedTotalPriceSold - calculatedTotalCost - shipping;
-
-      return {
-        ...updatedData,
-        //toFixed(2) ensures our calculations look like money. e.g $10.50
-        total_cost: Number(calculatedTotalCost.toFixed(2)),
-        total_price_sold: Number(calculatedTotalPriceSold.toFixed(2)),
-        profit: Number(calculatedProfit.toFixed(2)),
-      };
-    });
   };
 
   // Called when the user double-clicks a cell. We record which cell they clicked
@@ -272,7 +250,7 @@ export function InventoryForm({
                 if (field.type === "number") {
                   val = val.replace(/[^0-9.]/g, "");
                 }
-                handleChange(field.formName, val);
+                updateFormFields(field.formName, val, setFormData);
               }}
               placeholder={field.placeholder}
               readOnly={["total_cost", "total_price_sold", "profit"].includes(
